revno: 8721 message:   Change value-type of result in case aggregate query builder from Double to Integer.
committer: Tran Chau <tran.hispvietnam@gmail.com>
branch nick: dhis2
timestamp: Fri 2012-10-26 10:05:57 +0700

+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/caseaggregation/CaseAggregationConditionService.java
@@ -63,7 +63,7 @@ public interface CaseAggregationConditionService
     Collection<CaseAggregationCondition> getCaseAggregationCondition( Collection<DataElement> dataElements );
 
     
-    Double parseConditition( CaseAggregationCondition aggregationCondition, OrganisationUnit orgunit, Period period );
+    Integer parseConditition( CaseAggregationCondition aggregationCondition, OrganisationUnit orgunit, Period period );
         
     Collection<PatientDataValue> getPatientDataValues( CaseAggregationCondition aggregationCondition, OrganisationUnit orgunit, Period period );
